fix: make `wake_robust_futex` atomic.
This fix requires a patch in `ostd::mm::io::VmWriter::atomic_update`
to allow the caller to flag the update for interruption by returning
`None` on the `op` closure after performing arbitrary logic on the
value that was read.

The return type for `atomic_update` was enhanced to better capture
the diagnostics of what happened so the caller has the information
necessary to decide what to do, depending on the outcome.

This change affected the current consumer of this function,
`CurrentUserSpace::atomic_update`, which now breaks out of the loop
when the update was canceled. However, the caller of the latter
`atomic_update` was patched to always provide `Some(_)` for the new
value so the behavior didn't change.

Finally, `wake_robust_futex` was changed to use the upgraded
`VmWriter::atomic_update` and solve the atomicity problem. The
operation is canceled when the value read indicates that the futex
is no longer held by the current thread.

This patch also adds `reader_checked` and `writer_checked` to the
implementations of `CurrentUserSpace`, which are called by
`wake_robust_futex` and also by other (slightly refactored)
functions of `CurrentUserSpace`.